Rename "The Ride" menu item to "Rides" and change the url space to /rides
As we are now advertising multiple rides we are switching to /rides and this
page now includes information on multiple rides (although it is still static
and does not draw ride information from the db as we would ultimately like
it to do).

This also includes a Mailchimp signup form to register interest for the
upcoming Paris to London ride.

Signed-off-by: Michael Willmott <michael@knodium.com>