Full rewrite to ReactJS client-side app and DRF API backend
This is a very big rewrite of the entire site which moves away from the Django MVC model to a Django Rest Framework API backend and a ReactJS client side app. This is partly because i wanted to experiment with building such a site but the techbikers codebase needed a bit of tidying in order to make it clearer what was going on for any new developers that are going to start contributing to it. This new app has very clear dependencies and a simple REST API which the client app can be built and extended upon.

The app as it stands in this commit has actually be rolled back a bit in terms of functionality and has essentially implemented a read-only version of the site (login and authentication are working though). The reasoning is that the entire registration process needs to change anyway in order to properly support pre-registration waiting lists. Functionality will be added back in further commits as the required changes to the core structure of the app are made.

As part of this rewrite, Django has been updated to the latest version. The main thing to notice here is that South is no longer an installed app and is not needed to manage migrations of data models. As we are moving all the models into one Django app (core) instead of the various different apps (such as riders, rides, chapters etc), it does not matter so much that we are removing all the migration history with this commit. Future changes in the data model will be tracked using Django's own migrate features (which are essentially the same as South).

Techbikers is now a client side app with no rendering being conducted on the server. It is built using React to create components that make up the views and flux architecture to handle one way data flow. To help implement the flux architecture i'm making use of a fairly new library called Marty.js. This just handles a lot of the boilerplate needed to implement a version of flux.

User accounts are still exactly the same but authentication has changed quite a lot. We have moved away from session based authentication to token based authentication. This is very suitable for client side apps, secure and easy to implement. It should make it easy to add social login as well (something that is long overdue!)

Signed-off-by: Michael Willmott <michael@knodium.com>