Automatically create/delete a rider profile when a user is created/deleted
RiderProfile.create_profile_for_user(): When a user registers, we create a User model for them, but not a related RiderProfile model. The @receiver decorator solves this problem. Any time a User model is created, RiderProfile is asked to create a corresponding instance of itself. This ensures that we never have any orphan Users (that is to say, no Users that don't have an associated RiderProfile).

RiderProfile.delete_profile_for_user(): Complementary to RiderProfile.create_profile_for_user(), this method performs cleanup when we delete a User. When a User is deleted, the @reciever decorator is triggered, we look up the related RiderProfile object and delete it.

Signed-off-by: Michael Willmott <michael@knodium.com>