Introduce membership of a chapter
We have decided to introduce a membership fee for Techbikers. This will be done on a per chapter basis (although this isn't exposed to the user at the moment), so riders can become members of a chapter.

## Membership model
This introduces a new membership model into the system which links a user with a chapter. Multiple records can exist to show a user extending their membership (the period of a membership doesn't change, rather they just get more records which extend their membership). In a similar way to ride registrations, we add a new API endpoint which allows us to get a list of riders who are a member of a chapter as well as joining an authenticated user to a chapter and creating a Stripe charge (and associated sale record) if required.

## Changes to user model
We add a new property to the user model, which is exposed in the API, which tells us if the user is a member or not. It doesn't tell us which chapter they are a member of at the moment.

Signed-off-by: Michael Willmott <michael@knodium.com>