Make the charge method on the sales model generic
As we now want to be able to charge for things other than just a ride fee we want to make the charge method generic rather than linked to a specific object type. We now explicitly pass through the parameters needed to create a charge with Stripe and then create an associated sale record.

Signed-off-by: Michael Willmott <michael@knodium.com>