Change RideRider model to have pending riders
This changes the RideRider model (which is used to link a rider to a ride) to remove the pending field and introduce a state field. This determines whether the rider has yet been accepted onto the ride or not. This allows us to invite people to pre-register for a ride and then select which users we want to invite to complete the signup process and join the ride.

This makes changes to the APIs which list riders of a ride or rides a user is registered for. We only consider registrations that are confirmed (status of registered). This is also the case when calculating the remaining number of places on the ride.

There are two steps to the migration process for this model. The first makes all the required changes to the model but as we are creating a new status field for each existing record, we set the value of this field to registered (REG). The second migration changes the default value of status to be pending (PEN) which will take effect for any new records.

Signed-off-by: Michael Willmott <michael@knodium.com>