Add API endpoints for allowing users to register for a ride
This adds two new ride API endpoints which allow us to register users for a ride. The first step is pre-registration where we register the user for the ride with the status: pending. This is done by making an authenticated POST request to /api/rides/<id>/riders. When making a GET request to the same endpoint we return a list of the riders (using the rider model rather than the registration/riderider model). This requires us to use different serializers depending on what type of request is being made.

The second API is completing registration for a ride and charging the user (if required). This is done by making an authenticated POST request to /api/rides/<id>/riders/<riderId>/charge. A Stripe token must be provided and is really the only parameter required for this as we have the rideId and riderId as URL parameters (the riderId must be the same as the authenticated rider). Due to a slight quirk we do require that the rideId is sent in the post as well even though we don't do anything with it. This is due to the way the serialised behaves with a UNIQUE_TOGETHER constraint (which stop s a rider having more than one registration record for any given ride).

We introduce a new API permission which can check to see if a rider has been accepted onto a ride. We use this to make sure people don't attempt to try and complete their registration without having been invited to.

Signed-off-by: Michael Willmott <michael@knodium.com>