Ride preregistration
This adds a new interface which displays the ride registration state (if any) for an authenticated. It also allows this user the ability to pre-register for a ride. At this stage we collect a little more information off of the rider, some is used to update their profile (like the biography and statement of why they are doing Techbikers). The rest of the information is added to the registration payload.

The next steps required for completing registration are yet to be done. This does however lay the groundwork for the next steps by creating new payment components and it contains all the right state logic within the main registration component.

This also introduces the material design icons from Google which are used as part of showing the user what steps there are (and which one they are currently on).

Signed-off-by: Michael Willmott <michael@knodium.com>