Add partial users to the auth store
This introduces the concept of a partial user. This is information we have about a user before they are registered or logged in. This is passed to the app through querystring parameters. When the app loads we check for these parameters, add any values to the partialUser state in the authStore and then strip these parameters from the URL. When the anonymous user then goes to the signup form, we will populate this form with any information we already have about the user.

The reason for this is we have a collection of people on our mailing list with basic information about them. When we email them and invite them to register, we don't want to have to ask them for the same information again so this is a way of passing this information into the app through a link in an email sent to them.

Signed-off-by: Michael Willmott <michael@knodium.com>