Added a progress button
This can be used as a normal button where you pass through the onClick function as a prop. You can then set the state of this button to be loading, success or error. This is useful for forms where you want to stop people from clicking the button again should the request be slow.

Signed-off-by: Michael Willmott <michael@knodium.com>