Make sure URLs have a protocol before submitting to the server in signup
There is a stupid bug where URLs appear valid (and sort of are) on the client side when a user enters it in the website field in the signup process. However, the validation checks on the server are much stricter and require a protocol. If this isn't provided then the registration will silently fail. This fixes this!

Signed-off-by: Michael Willmott <michael@knodium.com>