Update signup form to include better error handling and check for existing users
We now check for a user account with the email address and throw a validation error if the account exists. We now handle errors by adding an error object to the riders store (which we clear when the route changes). This is fed back into the signup form so if the signup fails we can show the user what is going on.

Signed-off-by: Michael Willmott <michael@knodium.com>