Add full_cost column to rider model
As well as the ride price we can now specify the full cost of the ride (per rider). This allows us to provide more information to the user, especially in the context of "pay what you want". A lot of riders don't understand that the price they pay for a ride is subsidised by our sponsors. The price is the minimum they can contribute towards the ride but we will allow riders to cover the full cost (or more!) if they want.

Signed-off-by: Michael Willmott <michael@knodium.com>