Add fundraiser model and add fundraising fields to ride model
This adds a fundraiser model so each user who is going a ride (or technically even a user who isn't) can have a fundraising record associated with that ride. It is minimal at the moment and specific to the Just Giving API. It contains all the information we need to make further API calls to get information on the fundraiser as well as direct users to the fundraising page.

Note: the signOnUrl field is given to us when the fundraising page is created, it is a link the user can use to go to their new page and be logged in (as they will have already provided us with credentials). This link is only active for 20 minutes after the page is created.
Signed-off-by: Michael Willmott <michael@knodium.com>