Add API endpoint and UI to create fundraising page
This adds an endpoint which when called, creates a fundraising page with Just Giving using the user and ride details. It then creates a fundraising record on our end so we have the details of that page to use on our site.

If a user has registered for a ride and does not have a fundraising page setup then we will show a simple form to get that setup on the ride details page.

Signed-off-by: Michael Willmott <michael@knodium.com>