Add everything needed for authenticating using OAuth backends
This adds all the backend views and endpoints along with the front-end components needed for authenticating against external backends using OAuth. It doesn't actually implement this at the moment but everything needed is there for Facebook login (as the first one, it is very easy to configure others). This heavily uses the brilliant python-social library. The custom views and APIs are to make sure we use our token authentication rather than the default session based logins. This implements an API which takes an OAuth code and backend and exchanges it for an auth token.

The access token obtained is stored in the db agains the existing or new user record.

We must only use providers that require the user to have validated their address as we will login an existing user if their email address matches the email we get from the provider (and then we'll link the social account to their techbikers account).

Signed-off-by: Michael Willmott <michael@knodium.com>