Authenticate social login requests using any existing JWT
If the user is already logged in, send the JWT in the header of the social auth request. This will then associate the social account with the logged in user. If the user isn't logged in then it will default to associating with an account with the same email or creating a new account.

This also adds an authSuccess callback so that you can trigger an action once the social auth process has completed successfully.

Signed-off-by: Michael Willmott <michael@knodium.com>