Update fundraising model
This updates the fundraising model to include more details. It also includes a method to fetch details from Just Giving. This allows us to call the fetch_details method on a fundraising record which will in turn call the Just Giving API to get the latest fundraising details. If successful, it will update the current record with the latest details.

Signed-off-by: Michael Willmott <michael@knodium.com>