Calculate the total money raised per ride from fundraising records
Rather than making a call to the Just Giving API to get the total amount furnished, we calculate this from our fundraising records (which in turn are updated from Just Giving). This gives us a more accurate total as not all our fundraisers are linked to the event. Ideally the methods would produce the same result (and if all our fundraisers were linked to the event then they would), this is just a more flexible way of doing it.

Note: this doesn't account for different currencies so isn't actually 100% accurate. We need to fix this by making sure that all fundraisers list store the amount in the currency of the ride they are linked to, even if the rider is fundraising in a different currency.
Signed-off-by: Michael Willmott <michael@knodium.com>