Major rewrite of front-end app.
The previous front-end app was written when i really didn't know much about React and writing front-end apps. The ecosystem has progressed a lot in the last year and part of this has involved the core state management library we used (marty.js) no longer being actively maintained in favour of using Redux.

This app is pretty much a rewrite. The display components have been broken out and containers introduced. We're now using Redux for state management which is a lot better.

There are some regressions as of this commit in terms of functionality but we're getting there. This should make it a lot easier to continue developing (and hopefully with less bugs!)

Signed-off-by: Michael Willmott <mwillmott@gmail.com>