Store the page meta information in the main store
Rather than having static meta tags in the Head we now generate the tags using the MetaTags component and insert them into the Body (which isn't perfect but the best way to achieve what we need right now). If the page meta info changes in the state then the tags update to reflect this. This allows us to have different meta information based on the different rides.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>