Sync our users with our Mailchimp list
This creates a job that runs every two hours to update our Mailchimp list with the latest members. It adds them to groups based on the rides they are registered for (past and present). In order to do this we have a new field for the ride model so we know what Mailchimp group that ride is associated with.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>