Create “app” module in the client app
This is the beginning of restructuring the client app to split it out into different modules for the site. This should make it clearer what is going on and how the app works. It keeps everything related to specific features and sections of the site contained within one module rather than code spread across the codebase.

This is the initial module that contains the base level components and logic. More modules to follow!

Signed-off-by: Michael Willmott <mwillmott@gmail.com>