Add authentication endpoints for Auth0
We are preparing to migrate across to using Auth0 for authentication. In order to help with this, the database connection in Auth0 can check our API to validate existing user credentials and to check if users exist (by email).

See https://auth0.com/docs/connections/database/migrating for more information.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>