Password is not required when creating a local user
The password field is no longer required when creating a local user. This is because authentication is handled by Auth0 so we don’t care about passwords. A password might not even exist if the user logs in using a different provider (e.g. Facebook). Not setting a password will mean that people can’t actually authenticate against the local account (not that there are any endpoints to do that anymore).

Signed-off-by: Michael Willmott <mwillmott@gmail.com>