Add processing state for ride registration and use the Button component
This updates the app state to reflect that a user is currently registering. The Button component is passed this state so it will show as loading once a user has clicked to complete registration. This is important as it will stop a user clicking the button twice during payment (which could end up charging the user twice).

Signed-off-by: Michael Willmott <mwillmott@gmail.com>