Add an Errors module to capture and display errors
This allows us to store errors (with a type and message) in the app state. Each error has a key of our choosing. You can then display that error (if it exists) using the Errors component and passing the key through as the prop. It’s very simple but works and can be expanded as needed. Errors are cleared every time the user navigates to a new page.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>