Reduce the number of API calls when updating fundraising stats
The cron job that is used to update the fundraising stats previously made a call to the Just Giving API for each live fundraiser. This results in a lot of API calls over the course of a day. We can instead bulk fetch the fundraising pages by their event (ride) and then loop through the results from Just Giving and update the corresponding local fundraisers. This should drastically reduce the number of API calls made.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>