Account for manually created fundraising pages
Sometimes people create a fundraising page themselves and these then need linking to their Techbikers account. This is possible but the pages don’t automatically update after the changes made to make the process of updating much more efficient. This explicitly identifies fundraisers that were automatically updated so we can now include them in the cronjob which updates the fundraising stats.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>