Only return an array of user id’s in rides list API response
The call to this API was super slow because we were pulling in way more information than needed. It was loading the entire list of riders for each ride, their full profiles as well as all their fundraisers. This is much more information than we need to display the rides page and we can get all the other information from other existing API endpoints.

Signed-off-by: Michael Willmott <mwillmott@gmail.com>