This is a bit of trickery to migrate across to JSON fields supported by MySQL
We needed to migrate the data and ensure is was encoded properly. To do this we essentially turned the payload field into a TextField (as the data was being stored as LONGTEXT). We could then rename the column, create a new payload column of type JSON (supported by MySQL) and then serialise the json object from the string and save it to the new column. Then delete the old column.

Took ages to get this to work but it's quite a nice migration now!